pyproject.toml
[project]
name = "knot-mosaics"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Noah West", email = "westnoahb@gmail.com"},
]
dependencies = ["pillow>=12.1.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
